requirements.txt
# ==============================================
# Umrah Planner AI - Requirements
# Python 3.10+ Required
# Optimized for Railway.com Deployment
# ==============================================

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM & AI - Groq & OpenAI Support
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0
langchain-openai>=0.0.2
openai>=1.6.0
groq>=0.4.0

# RAG Components
chromadb>=0.4.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
httpx>=0.25.0
tenacity>=8.2.0
rich>=13.0.0

# Export & Visualization
plotly>=5.18.0
openpyxl>=3.1.0
fpdf2>=2.7.0
reportlab>=4.0.0

# Production Server (Railway)
gunicorn>=21.0.0
watchdog>=3.0.0

# Note: faiss-cpu removed for Railway compatibility
# ChromaDB is sufficient for this use case
